Add first version of the check_named.rb Nagios NRPE check.
This check allows to deterime whether named from the BIND software
suite is still performing its duties as it should and did not ended
up in rather odd state where it refuses to answer queries after
short networking outage.  This only happens where forwarders are
involved and it is very annoying.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>